refactor(error): Split files
